alpha.3 verification: fix pre-existing clippy + test-compile warts
Systematic verification of alpha.3's distributed restoration surfaced
three pre-existing hygiene issues (all present at v0.2.0-beta.5, not
alpha.3 regressions), fixed as drive-by so `cargo clippy --workspace
--all-targets -- -D warnings` passes cleanly.

- lance-admin: three `o.size as u64` casts became no-ops after
  object_store updated ObjectMeta::size to u64. clippy::unnecessary_cast.
- lance-integration integration tests: seven files constructed
  AnnSearchRequest as a struct literal without min_schema_version
  and min_commit_seq (the #5.4/#5.5 fields added in v2). Integration
  tests have been broken since those proto fields landed; now fixed.

Verification results
  Distributed capability (A): scatter_gather + shard_pruning +
    partition_batch_by_hash all live, byte-identical to v0.2.0-beta.5.
    CreateTable auto-sharding, MoveShard, rebalance, connection-pool
    shard recovery all active.
  New R3 capabilities (B): AlterTable/CreateTag/DeleteTag/
    RestoreTable fan-out across every shard under DDL lease.
    ListTags picks any healthy shard (converged set). CreateIndex
    dispatches all supported types correctly (bug fix preserved).
  Code health (C): workspace lib clippy clean, all-targets clippy
    clean after drive-by fixes. 541 lib tests pass. No dead-code
    warnings in the coord.
  Tests (D): e2e_tags 22/22, e2e_schema_evolution 18/18, e2e_commit_seq
    12/12, full E2E 13/13 in a clean run. Flakes observed only
    under port-contention after prior runs; pass in isolation.

+++ crates/admin/src/main.rs
@@ -521,7 +521,7 @@ async fn cmd_shards_copy(
         }
 
         if dry_run {
-            let bytes: u64 = objects.iter().map(|o| o.size as u64).sum();
+            let bytes: u64 = objects.iter().map(|o| o.size).sum();
             println!("{table}\t{rel}\t{} objects\t{} bytes", objects.len(), bytes);
             total_objects += objects.len() as u64;
             total_bytes += bytes;
@@ -546,7 +546,7 @@ async fn cmd_shards_copy(
                 dst_store.as_ref(),
                 &o.location,
                 &dst_path,
-                o.size as u64,
+                o.size,
                 multipart_threshold,
             ).await;
 
@@ -554,7 +554,7 @@ async fn cmd_shards_copy(
                 Ok(()) => {
                     log::info!("copy {} → {} ({} bytes)", o.location, dst_path, o.size);
                     total_objects += 1;
-                    total_bytes += o.size as u64;
+                    total_bytes += o.size;
                 }
                 Err(e) => {
                     failures.push((o.location.to_string(), e.to_string()));
